+++ src/datagen/java/net/frozenblock/wilderwild/datagen/tag/WWBiomeTagProvider.java
@@ -1395,6 +1395,9 @@ public final class WWBiomeTagProvider extends FrozenBiomeTagProvider {
 		this.getOrCreateTagBuilder(WilderBiomeTags.HAS_TERMITE_MOUND)
 			.addOptionalTag(BiomeTags.IS_SAVANNA);
 
+		this.getOrCreateTagBuilder(WilderBiomeTags.HAS_NETHER_GEYSER)
+			.add(Biomes.BASALT_DELTAS);
+
 		this.getOrCreateTagBuilder(WilderBiomeTags.HAS_TAIGA_FOREST_ROCK)
 			.add(Biomes.TAIGA)
 			.add(Biomes.SNOWY_TAIGA)