+++ src/main/java/net/frozenblock/wilderwild/config/WorldgenConfig.java
@@ -92,6 +92,9 @@ public final class WorldgenConfig {
 	@EntrySyncData("termiteGen")
 	public boolean termiteGen = true;
 
+	@EntrySyncData("basaltDeltasGeyserGen")
+	public boolean basaltDeltasGeyserGen = true;
+
 	@EntrySyncData("surfaceDecoration")
 	public boolean surfaceDecoration = true;